chore: move batcher to live-network
