+++ libs/live-network/src/index.ts
@@ -1,0 +1,2 @@
+import network from "./network";
+export default network;
