+++ libs/live-network/src/network.ts
@@ -1,10 +1,15 @@
+import axios, {
+  AxiosPromise,
+  AxiosResponse,
+  type AxiosError,
+  type AxiosRequestConfig,
+  type Method,
+} from "axios";
+import invariant from "invariant";
 import { LedgerAPI4xx, LedgerAPI5xx, NetworkDown } from "@ledgerhq/errors";
 import { changes, getEnv } from "@ledgerhq/live-env";
 import { retry } from "@ledgerhq/live-promise";
 import { log } from "@ledgerhq/logs";
-import type { AxiosError, AxiosRequestConfig, Method } from "axios";
-import axios, { AxiosPromise, AxiosResponse } from "axios";
-import invariant from "invariant";
 
 type Metadata = { startTime: number };
 type ExtendedXHRConfig = AxiosRequestConfig & { metadata?: Metadata };
